Profile

+++ client/src/components/containers/Profile.js
@@ -14,25 +14,29 @@ const Profile = props => {
   const [updateMode, setUpdateMode] = useState(false);
 
   useEffect( () => {
+	getReservations();
+  }, []);
+
+  const getReservations = () => {
 	axios.get('/api/user/'+authContext.user.userID)
 		 .then( res => {
 		   const responseData = res.data.reservations;
-		   const updatedReservations = [...reservations];
+		   const updatedReservations = [];
 		   for (let i = 0; i < responseData.length; i++) {
 			 updatedReservations.push(responseData[i].row.slice(1,-1).split(','));
 		   };
 		   setReservations(updatedReservations)
 		 })
 		 .catch( err => console.log('Err: ', err));
-  }, []);
+  }
 
 
   const cancelReservation = resId => {
-	console.log('E: ', resId);
-	axios.delete('/api/reservation/'+resId)
-		 .then( data => console.log('Rows deleted: ', data))
-		 .catch( err => console.log('Error: ', err));
-
+	if (confirm('Are you sure you want to cancel this reservation?')) {
+	  axios.delete('/api/reservation/'+resId)
+		   .then( res => getReservations() )
+		   .catch( err => console.log('Error: ', err));
+	};
   };
 
   const active = reservations.filter( (res, i) => {
@@ -40,7 +44,7 @@ const Profile = props => {
   });
 
   const currentReservations = active.map( (res, i) => (
-    <li key={i}>
+    <li key={i} className='reservation'>
 	  {moment(res[0]).format('ddd MMM DD YYYY HH:MM')} : {res[2]} for {moment.duration(moment(res[1]).diff(moment(res[0]))).asHours().toFixed(1)} hours
 	  <IconButton className='cancel-button' size='small' data-id={res.id} onClick={() => cancelReservation(res[4])} aria-label='Cancel this reservation'> <Cancel fontSize='inherit'/> </IconButton>
 	</li>
@@ -51,7 +55,9 @@ const Profile = props => {
   });
 
   const pastReservations = done.map( (res, i) => (
-    <li key={i}>{moment(res[0]).format('ddd MMM DD YYYY HH:MM')} : {res[2]} for {moment.duration(moment(res[1]).diff(moment(res[0]))).asHours().toFixed(1)} hours</li>
+    <li key={i} className='reservation'>
+	  {moment(res[0]).format('ddd MMM DD YYYY HH:MM')} : {res[2]} for {moment.duration(moment(res[1]).diff(moment(res[0]))).asHours().toFixed(1)} hours
+	</li>
   ));
 
   const toggleMode = () => setUpdateMode(!updateMode);
@@ -65,7 +71,7 @@ const Profile = props => {
 		<InputLabel id='phone-input'>Phone Number</InputLabel>
 		<TextField placeholder={authContext.user.phone} className='input' fullWidth variant='outlined' name='first_name' type='text' />
 		<Button id='update-user' className='auth-button' color='primary' variant='contained' onClick={() => console.log('TODO: create save handler')} startIcon={<Save />}>Save Profile</Button>
-		<Button id='update-user' className='auth-button' color='default' variant='contained' onClick={toggleMode} startIcon={<Cancel />}>Cancel</Button>
+		<Button id='update-user' className='auth-button' color='secondary' variant='contained' onClick={toggleMode} startIcon={<Cancel />}>Cancel</Button>
 	  </form>
   );
